MEDI-50 Adding Controllers
